Consume @opena2a/atx-verify in broker; remove bundled verifier copy
The ATX verifier was maintained as src/broker/atx.ts here and again in the
AIM SDK — two copies of byte-sensitive JCS canonicalization that had to track
the Go/Python reference verifiers in lockstep. Swap the broker to the shared,
separately-tested @opena2a/atx-verify package and delete the local copy,
closing the drift window.

- All ATX imports across the broker (grant-resolver, grant-policy, cpi/*) and
  the barrels repoint to the package. The package is ESM-only and secretless is
  CommonJS, so type-only sites use 'resolution-mode: import' (the established
  cli-ui/telemetry pattern); no runtime require() of the ESM package is emitted.
- Public surface: the verifier VALUES (LocalAtxVerifier, canonicalPayload,
  normalizeRfc3339, SUPPORTED_ATX_VERSION) are no longer re-exported (CJS can't
  re-export ESM values) — import them from @opena2a/atx-verify. ATX types stay
  re-exported. CLI behavior unchanged.
- AAP conformance test now sources its fixtures inline (was reusing helpers
  exported from the deleted atx.test.ts) and instantiates LocalAtxVerifier from
  the package. Remove canonicalize as a direct dep (now transitive).
