+++ src/broker/index.ts
@@ -32,20 +32,25 @@ export {
 } from './events';
 
 // AAP (Agent Authorization Protocol) — grant-based authorization layer.
-export {
-  LocalAtxVerifier,
-  canonicalPayload,
-  normalizeRfc3339,
-  SUPPORTED_ATX_VERSION,
-  type Atx,
-  type AtxSignature,
-  type AtxPublicKey,
-  type AtxTrustAnchors,
-  type AtxVerifier,
-  type AtxVerificationResult,
-  type ResolutionContext,
-  type RejectCategory,
-} from './atx';
+//
+// The ATX verifier was extracted to @opena2a/atx-verify (the single
+// spec-compliant implementation, shared with the AIM SDK). secretless
+// re-exports its TYPES only: types are erased at compile, so they cross the
+// CJS→ESM-only package boundary for free, and secretless's own public surface
+// (GrantResolverDeps.verifier: AtxVerifier, GrantResolveInput.atx: Atx) needs
+// them. Callers wanting the VALUES (LocalAtxVerifier, canonicalPayload,
+// normalizeRfc3339, SUPPORTED_ATX_VERSION) import them from @opena2a/atx-verify
+// directly — a CJS package cannot statically re-export ESM-only values.
+export type {
+  Atx,
+  AtxSignature,
+  AtxPublicKey,
+  AtxTrustAnchors,
+  AtxVerifier,
+  AtxVerificationResult,
+  ResolutionContext,
+  RejectCategory,
+} from '@opena2a/atx-verify' with { 'resolution-mode': 'import' };
 export {
   GrantPolicy,
   type GrantBinding,
